Auth, Teaching Load, Teaching Load Details, Enrollment, need Read, Update and Delete

+++ src/main/java/com/assookkaa/ClassRecord/Dto/Request/GradingComposition/GradingCompositionDtoRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class GradingCompositionDtoRequest {
-    private double percentage;
+    private Float percentage;
     private Integer categoryId;
     private Integer teachingLoadDetailId;
 }
